Added code to properly handle covariance of input variables (closes issue #4)
Added code to consistently draw from the input arrays such that covariance is
properly accounted for. Currently if the M, C, and D_proj arrays are the same
size then it is assumed that the values and order of the input arrays are properly
correlated and during the random draw phase the same randomly drawn index
will be used to sample all arrays.