Adjusted v_3d_col > v_3d_max portion of code, closes #10
Now this recognizes the scenario as being representative of the unbound case and will save
the calculated parameters up to this point, ignoring those that have no meaning for the unbound
scenario. Note that this does assume that the two subclusters are the dominate systems in the
local area and cannot merge with a velocity greater than that of the free fall velocity.